--- ThreeXPlusOne/Code/Graph/GraphProviders/GraphProvider.cs
@@ -1,6 +1,0 @@
-namespace ThreeXPlusOne.Code.Graph.GraphProviders;
-
-public enum GraphProvider
-{
-    SkiaSharp = 1
-}
