+++ ThreeXPlusOne/Code/Graph/Services/GraphProvider.cs
@@ -1,0 +1,6 @@
+namespace ThreeXPlusOne.Code.Graph.Services;
+
+public enum GraphProvider
+{
+    SkiaSharp = 1
+}
