+++ ThreeXPlusOne/StartupExtensions.cs
@@ -3,7 +3,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using ThreeXPlusOne.Code;
 using ThreeXPlusOne.Code.Graph;
-using ThreeXPlusOne.Code.Graph.GraphProviders;
+using ThreeXPlusOne.Code.Graph.Services;
 using ThreeXPlusOne.Code.Helpers;
 using ThreeXPlusOne.Code.Interfaces;
 using ThreeXPlusOne.Config;
